refactor: Moved URL handling code to its own utils file
